Fixed Class Syntax
The problem with the class syntax was 1) a damn parentheses and 2) the
class definition needs to be sequentially before the method that it is
instantiated in. Probably be worth making a library for all the classes
and import it in, just to avoid this kind of problem.

Also added user input to chose path name based on location.